requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
authlib==1.6.5
bcrypt==5.0.0
beartype==0.22.5
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cyclopts==4.2.4
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.3
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.121.2
fastmcp==2.13.1
fastuuid==0.14.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
gitpython==3.1.45
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.1.4
idna==3.11
importlib-metadata==8.7.0
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain-core==0.3.79
langchain-litellm==0.3.0
langchain-openai
langchain-mcp-adapters==0.1.13
langsmith==0.4.43
litellm==1.80.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.21.2
mdurl==0.1.2
more-itertools==10.8.0
multidict==6.7.0
narwhals==2.12.0
numpy==2.3.5
openai==2.8.0
openapi-pydantic==0.5.1
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pathable==0.4.4
pathvalidate==3.3.1
pillow==12.0.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.1
psycopg2-binary==2.9.11
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydeck==0.9.1
pygments==2.19.2
pyjwt==2.10.1
pymilvus==2.6.3
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.29.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
